+++ src/net.cpp
@@ -2951,8 +2951,7 @@ void CNode::AskFor(const CInv &inv) {
         nRequestTime = 0;
     }
     LogPrint(BCLog::NET, "askfor %s  %d (%s) peer=%d\n", inv.ToString(),
-             nRequestTime,
-             DateTimeStrFormat("%H:%M:%S", nRequestTime / 1000000), id);
+             nRequestTime, FormatISO8601DateTime(nRequestTime / 1000000), id);
 
     // Make sure not to reuse time indexes to keep things in the same order
     int64_t nNow = GetTimeMicros() - 1000000;